[IMCC] Fixed a compiler warning.
I fear this will come back when someone regenerates the file.  Hateful lex.

git-svn-id: https://svn.parrot.org/parrot/trunk@18653 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ compilers/imcc/imclexer.c
@@ -2879,7 +2879,7 @@ static int input (yyscan_t yyscanner );
 		{ \
 		int c = '*'; \
 		size_t n; \
-		for ( n = 0; n < max_size && \
+		for ( n = 0; n < (size_t)max_size && \
 			     (c = getc( yyin )) != EOF && c != '\n'; ++n ) \
 			buf[n] = (char) c; \
 		if ( c == '\n' ) \
